pyproject.toml
[tool.poetry]
name = "reports-runner"
version = "24.0.0"
description = "Connect Reports Runner"
authors = ["CloudBlue"]
license = "Apache-2.0"
packages = [
    { include = "executor" }
]
readme = "./README.md"
documentation = "https://confluence.int.zone/display/CONNECT/Reports+Management"
homepage = "https://connect.cloudblue.com"
repository = "https://git.int.zone/projects/SWFT/repos/reports-runner"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.8",
    "Topic :: Utilities",
]
keywords = [
    "utility",
    "connect",
    "cloudblue",
    "reports",
]

[tool.poetry.scripts]
cextrun = 'executor.runner:run_executor'

[tool.poetry.dependencies]
chardet = "3.*"
connect-openapi-client = ">=22.0.8"
connect-reports-core = "23.0.*"
lxml = "4.*"
openpyxl = "3.*"
python = "3.8.*"
requests = "2.*"

[tool.poetry.dev-dependencies]
flake8 = "3.*"
flake8-broken-line = "0.*"
flake8-bugbear = "21.*"
flake8-cognitive-complexity = "0.*"
flake8-commas = "2.*"
flake8-comprehensions = "3.*"
flake8-debugger = "3.*"
flake8-eradicate = "1.*"
flake8-future-import = "0.*"
flake8-import-order = "0.*"
flake8-string-format = "0.*"
ipdb = "0.*"
ipython = "7.*"
jsonschema = "3.*"
pdbpp = "0.*"
pyfakefs ="4.*"
pytest = "6.*"
pytest-cov = "2.*"
pytest-deadfixtures = "2.*"
pytest-mock = "3.*"
pytest-randomly = "3.*"
responses = "0.*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
addopts = "--cov=executor --cov-report=term-missing:skip-covered --cov-report=html --cov-report=xml"

[tool.coverage.run]
branch = true

[tool.coverage.report]
omit = [
    "*/migrations/*",
    "*/config/*",
    "*/settings/*",
    "*/manage.py",
    "*/wsgi.py",
    "*/urls.py"
]

exclude_lines = [
    "pragma: no cover",
    "def __str__",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]


[tool.flakehell]

exclude = [
    ".idea",
    ".vscode",
    ".git",
    "postgres_data",
    "*/migrations/*.py",
    "*/settings/*.py",
    "*/manage.py",
    "*.eggs",
    "*.egg",
    "tests/",
    "./docs/*.rst",
    "demo.ipynb"
]
show_source = true
max_line_length = 120
# application-import-names = "smarkets"

[tool.flakehell.plugins]
pycodestyle = ["+*", "-W503"]
pyflakes = ["+*"]
pylint = ["-*"]

[tool.isort]
line_length = 100
combine_as_imports = true
balanced_wrapping = true
lines_after_imports = 2
include_trailing_comma = true
not_skip = "__init__.py"
multi_line_output = 5

import_heading_stdlib = "built-in"
import_heading_thirdparty = "external"
import_heading_firstparty = "project"
